src/auth/requirements.txt
jwt
flask
flask
Flask-MySQLdb